requirements.txt
streamlit
faiss-cpu
sentence-transformers
openai
pypdf
numpy
scikit-learn
torch